Refactor treatment efficacy prediction to use ShapDetail and SimilarPatientsSummary data structures
