+++ common/src/main/kotlin/com/hartwig/actin/personalization/serialization/PersonalizedTreatmentSummaryJson.kt
@@ -14,10 +14,10 @@ object PersonalizedTreatmentSummaryJson {
     private val logger = LogManager.getLogger(PersonalizedTreatmentSummaryJson::class.java)
     private val mapper = ObjectMapper().registerModule(KotlinModule.Builder().build())
 
-    fun read(treatmentEfficacyPredictionJson: String): PersonalizedTreatmentSummary {
-        logger.info("Loading treatment efficacy predictions from $treatmentEfficacyPredictionJson")
+    fun read(personalizedTreatmentSummary: String): PersonalizedTreatmentSummary {
+        logger.info("Loading treatment efficacy predictions from $personalizedTreatmentSummary")
 
-        val json = Files.readString(File(treatmentEfficacyPredictionJson).toPath())
+        val json = Files.readString(File(personalizedTreatmentSummary).toPath())
         val results = mapper.readValue<PersonalizedTreatmentSummary>(json)
 
         logger.info(" Loaded treatment efficacy predictions for ${results.predictions?.size} treatments")
