+++ datamodel/src/main/kotlin/com/hartwig/actin/datamodel/algo/ShapDetail.kt
@@ -1,0 +1,6 @@
+package com.hartwig.actin.datamodel.algo
+
+data class ShapDetail(
+    val featureValue: Double,
+    val shapValue: Double
+)
