+++ datamodel/src/main/kotlin/com/hartwig/actin/datamodel/algo/TreatmentMatch.kt
@@ -17,24 +17,4 @@ data class TreatmentMatch(
 data class PersonalizedTreatmentSummary(
     val predictions: List<TreatmentEfficacyPrediction>?,
     val similarPatientsSummary: SimilarPatientsSummary?
-) {
-    data class TreatmentEfficacyPrediction(
-        val treatment: String,
-        val survivalProbs: List<Double>,
-        val shapValues: Map<String, ShapDetail>
-    ) {
-        data class ShapDetail(
-            val featureValue: Double,
-            val shapValue: Double
-        )
-    }
-    data class SimilarPatientsSummary(
-        val overallTreatmentProportion: List<TreatmentProportion>,
-        val similarPatientsTreatmentProportion: List<TreatmentProportion>
-    ) {
-        data class TreatmentProportion(
-            val treatment: String,
-            val proportion: Double
-        )
-    }
-}
+)
