+++ datamodel/src/main/kotlin/com/hartwig/actin/datamodel/algo/TreatmentProportion.kt
@@ -1,0 +1,6 @@
+package com.hartwig.actin.datamodel.algo
+
+data class TreatmentProportion(
+    val treatment: String,
+    val proportion: Double
+)
